исправил замечания по отображению карточек
